+++ CSK_Module_MultiRemoteCamera/scripts/Sensors/MultiRemoteCamera/helper/funcs.lua
@@ -7,9 +7,13 @@
 --**********************Start Global Scope *********************************
 --**************************************************************************
 
+local nameOfModule = 'CSK_MultiRemoteCamera'
+
 local funcs = {}
 -- Providing standard JSON functions
 funcs.json = require('Sensors/MultiRemoteCamera/helper/Json')
+-- Default parameters for instances of module
+funcs.defaultParameters = require('Sensors/MultiRemoteCamera/MultiRemoteCamera_Parameters')
 
 --**************************************************************************
 --********************** End Global Scope **********************************
@@ -132,6 +136,30 @@ local function convertContainer2Table(cont)
   return data
 end
 funcs.convertContainer2Table = convertContainer2Table
+
+--- Function to compare table content. Optionally will fill missing values within content table with values of defaultTable
+---@param content auto Data to check
+---@param defaultTable auto Reference data
+---@return auto[] content Update of data
+local function checkParameters(content, defaultTable)
+  for key, value in pairs(defaultTable) do
+    if type(value) == 'table' then
+      if content[key] == nil then
+        _G.logger:info(nameOfModule .. ": Created missing parameters table '" .. tostring(key) .. "'")
+        content[key] = {}
+      end
+      content[key] = checkParameters(content[key], defaultTable[key])
+    elseif content[key] == nil then
+      _G.logger:info(nameOfModule .. ": Missing parameter '" .. tostring(key) .. "'. Adding default value '" .. tostring(defaultTable[key]) .. "'")
+      content[key] = defaultTable[key]
+      if key == 'cameraNo' then
+        _G.logger:warning(nameOfModule .. ": '" .. tostring(key) .. "' is a major parameter! Default value might not work and needs to be edited!")
+      end
+    end
+  end
+  return content
+end
+funcs.checkParameters = checkParameters
 
 return funcs